IPVGO: refactor opponent list getter
